Resync from Contentful whenever user pulls down
